Fix build in aarch and unsupported

+++ cpp/Platform.Hashing.Tests/AllTests.cpp
@@ -9,10 +9,10 @@ namespace Platform::Hashing::Tests
 {
     TEST(CombineTest, Trivial)
     {
-        std::uint32_t hash1 = 0;
+        std::size_t hash1 = 0;
         Combine(hash1, 42);
         ASSERT_NE(0U, hash1);
-        std::uint32_t hash2 = 0;
+        std::size_t hash2 = 0;
         Combine(hash2, 42);
         ASSERT_EQ(hash2, hash1);
     }
